project/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.4.0
numpy==2.3.0
reportlab==4.0.7

# YOLOv8 Neutrophil Detection
ultralytics>=8.4.37
torch>=2.11.0
torchvision>=0.26.0
opencv-python>=4.13.0.92
PyYAML>=6.0.3
matplotlib>=3.10.8
scipy>=1.17.1
